# apps/dashboard/src/lib/openapi.json
{
  "components": {
    "schemas": {
...
      "DetectionStatusRequest": {
        "additionalProperties": false,
        "properties": {
          "status": {
            "enum": [
              "new",
              "acknowledged",
              "false_positive",
              "resolved"
            ],
            "title": "Status",
            "type": "string"
          }
        },
        "required": [
          "status"
        ],
        "title": "DetectionStatusRequest",
        "type": "object"
      },
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
      "IncidentStatus": {
        "description": "Lifecycle of a correlated incident.",
        "enum": [
          "open",
          "investigating",
          "contained",
          "resolved",
          "false_positive"
        ],
        "title": "IncidentStatus",
        "type": "string"
      },
      "IncidentUpdateRequest": {
        "additionalProperties": false,
        "properties": {
          "assigned_to": {
            "anyOf": [
              {
                "maxLength": 64,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Assigned To"
          },
          "notes": {
            "anyOf": [
              {
                "maxLength": 10000,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Notes"
          },
          "status": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/IncidentStatus"
              },
              {
                "type": "null"
              }
            ]
          }
        },
        "title": "IncidentUpdateRequest",
        "type": "object"
      },
...
      "Severity": {
        "description": "How serious a detection is, before risk scoring weighs the context.",
        "enum": [
          "info",
          "low",
          "medium",
          "high",
          "critical"
        ],
        "title": "Severity",
        "type": "string"
      },
      "ThreatCategory": {
        "description": "What kind of activity a detection represents.",
        "enum": [
          "reconnaissance",
          "brute_force",
          "denial_of_service",
          "exfiltration",
          "protocol_anomaly",
          "policy_violation",
          "malicious_reputation",
          "anomaly",
          "lateral_movement",
          "other"
        ],
        "title": "ThreatCategory",
        "type": "string"
      },
...
      "ValidationError": {
        "properties": {
          "ctx": {
            "title": "Context",
            "type": "object"
          },
          "input": {
            "title": "Input"
          },
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Location",
            "type": "array"
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        },
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError",
        "type": "object"
      }
    }
  },
  "info": {
    "description": "\nSentinelX network intrusion detection and prevention API.\n\n**Authentication.** Obtain tokens from `POST /api/v1/auth/login` and send\n`Authorization: Bearer <access_token>`. Browser sessions use httpOnly cookies and must\necho the `sx_csrf` cookie in an `X-CSRF-Token` header on state-changing requests.\n\n**Roles.** `viewer` reads; `analyst` triages, tests rules and runs replays; `admin`\nchanges rules, firewall state and settings.\n\n**Safety.** Prevention is off by default. Responses honour `DRY_RUN`, and every\nblock passes the safety guard whether it came from a detector or a person.\n",
    "title": "SentinelX API",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
    "/api/v1/alerts": {
      "get": {
        "operationId": "alerts_api_v1_alerts_get",
        "parameters": [
          {
            "in": "query",
            "name": "hours",
            "required": false,
            "schema": {
              "default": 24,
              "maximum": 720,
              "minimum": 1,
              "title": "Hours",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": true,
                  "title": "Response Alerts Api V1 Alerts Get",
                  "type": "object"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Items needing attention: high-risk untriaged detections and pending approvals",
        "tags": [
          "alerts"
        ]
      }
    },
...
    "/api/v1/detections": {
      "get": {
        "operationId": "list_detections_api_v1_detections_get",
        "parameters": [
          {
            "in": "query",
            "name": "severity",
            "required": false,
            "schema": {
              "default": [],
              "items": {
                "$ref": "#/components/schemas/Severity"
              },
              "title": "Severity",
              "type": "array"
            }
          },
          {
            "in": "query",
            "name": "detector",
            "required": false,
            "schema": {
              "default": [],
              "items": {
                "type": "string"
              },
              "maxItems": 50,
              "title": "Detector",
              "type": "array"
            }
          },
          {
            "in": "query",
            "name": "category",
            "required": false,
            "schema": {
              "default": [],
              "items": {
                "$ref": "#/components/schemas/ThreatCategory"
              },
              "title": "Category",
              "type": "array"
            }
          },
          {
            "in": "query",
            "name": "status",
            "required": false,
            "schema": {
              "default": [],
              "items": {
                "enum": [
                  "new",
                  "acknowledged",
                  "false_positive",
                  "resolved"
                ],
                "type": "string"
              },
              "title": "Status",
              "type": "array"
            }
          },
          {
            "in": "query",
            "name": "source_ip",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "maxLength": 64,
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Source Ip"
            }
          },
          {
            "in": "query",
            "name": "destination_ip",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "maxLength": 64,
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Destination Ip"
            }
          },
          {
            "in": "query",
            "name": "protocol",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "maxLength": 16,
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Protocol"
            }
          },
          {
            "in": "query",
            "name": "incident_id",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "maxLength": 64,
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Incident Id"
            }
          },
          {
            "in": "query",
            "name": "replay_id",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "maxLength": 64,
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Replay Id"
            }
          },
          {
            "in": "query",
            "name": "since",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "format": "date-time",
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Since"
            }
          },
          {
            "in": "query",
            "name": "until",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "format": "date-time",
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Until"
            }
          },
          {
            "in": "query",
            "name": "min_risk",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "maximum": 100,
                  "minimum": 0,
                  "type": "number"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Min Risk"
            }
          },
          {
            "description": "Free-text search",
            "in": "query",
            "name": "q",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "maxLength": 200,
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Free-text search",
              "title": "Q"
            }
          },
          {
            "in": "query",
            "name": "order",
            "required": false,
            "schema": {
              "default": "newest",
              "enum": [
                "newest",
                "oldest",
                "risk"
              ],
              "title": "Order",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "limit",
            "required": false,
            "schema": {
              "default": 50,
              "maximum": 500,
              "minimum": 1,
              "title": "Limit",
              "type": "integer"
            }
          },
          {
            "in": "query",
            "name": "offset",
            "required": false,
            "schema": {
              "default": 0,
              "minimum": 0,
              "title": "Offset",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": true,
                  "title": "Response List Detections Api V1 Detections Get",
                  "type": "object"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "List Detections",
        "tags": [
          "detections"
        ]
      }
    },
    "/api/v1/detections/{detection_id}": {
      "get": {
        "operationId": "get_detection_api_v1_detections__detection_id__get",
        "parameters": [
          {
            "in": "path",
            "name": "detection_id",
            "required": true,
            "schema": {
              "title": "Detection Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": true,
                  "title": "Response Get Detection Api V1 Detections  Detection Id  Get",
                  "type": "object"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Get Detection",
        "tags": [
          "detections"
        ]
      },
      "patch": {
        "operationId": "update_detection_api_v1_detections__detection_id__patch",
        "parameters": [
          {
            "in": "path",
            "name": "detection_id",
            "required": true,
            "schema": {
              "title": "Detection Id",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/DetectionStatusRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": true,
                  "title": "Response Update Detection Api V1 Detections  Detection Id  Patch",
                  "type": "object"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Triage: acknowledge, resolve or mark false positive",
        "tags": [
          "detections"
        ]
      }
    },
...
    "/api/v1/incidents": {
      "get": {
        "operationId": "list_incidents_api_v1_incidents_get",
        "parameters": [
          {
            "in": "query",
            "name": "status",
            "required": false,
            "schema": {
              "default": [],
              "items": {
                "$ref": "#/components/schemas/IncidentStatus"
              },
              "title": "Status",
              "type": "array"
            }
          },
          {
            "in": "query",
            "name": "severity",
            "required": false,
            "schema": {
              "default": [],
              "items": {
                "$ref": "#/components/schemas/Severity"
              },
              "title": "Severity",
              "type": "array"
            }
          },
          {
            "in": "query",
            "name": "min_risk",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "maximum": 100,
                  "minimum": 0,
                  "type": "number"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Min Risk"
            }
          },
          {
            "in": "query",
            "name": "replay_id",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "maxLength": 64,
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Replay Id"
            }
          },
          {
            "in": "query",
            "name": "since",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "format": "date-time",
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Since"
            }
          },
          {
            "in": "query",
            "name": "limit",
            "required": false,
            "schema": {
              "default": 50,
              "maximum": 500,
              "minimum": 1,
              "title": "Limit",
              "type": "integer"
            }
          },
          {
            "in": "query",
            "name": "offset",
            "required": false,
            "schema": {
              "default": 0,
              "minimum": 0,
              "title": "Offset",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": true,
                  "title": "Response List Incidents Api V1 Incidents Get",
                  "type": "object"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "List Incidents",
        "tags": [
          "incidents"
        ]
      }
    },
    "/api/v1/incidents/{incident_id}": {
      "get": {
        "operationId": "get_incident_api_v1_incidents__incident_id__get",
        "parameters": [
          {
            "in": "path",
            "name": "incident_id",
            "required": true,
            "schema": {
              "title": "Incident Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": true,
                  "title": "Response Get Incident Api V1 Incidents  Incident Id  Get",
                  "type": "object"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Get Incident",
        "tags": [
          "incidents"
        ]
      },
      "patch": {
        "operationId": "update_incident_api_v1_incidents__incident_id__patch",
        "parameters": [
          {
            "in": "path",
            "name": "incident_id",
            "required": true,
            "schema": {
              "title": "Incident Id",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/IncidentUpdateRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": true,
                  "title": "Response Update Incident Api V1 Incidents  Incident Id  Patch",
                  "type": "object"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Update Incident",
        "tags": [
          "incidents"
        ]
      }
    },
...
    "/api/v1/threats": {
      "get": {
        "operationId": "threats_api_v1_threats_get",
        "parameters": [
          {
            "in": "query",
            "name": "hours",
            "required": false,
            "schema": {
              "default": 24,
              "maximum": 720,
              "minimum": 1,
              "title": "Hours",
              "type": "integer"
            }
          },
          {
            "in": "query",
            "name": "limit",
            "required": false,
            "schema": {
              "default": 100,
              "maximum": 500,
              "minimum": 1,
              "title": "Limit",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "additionalProperties": true,
                    "type": "object"
                  },
                  "title": "Response Threats Api V1 Threats Get",
                  "type": "array"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Detections grouped by source address",
        "tags": [
          "threats"
        ]
      }
    },
...
  }
}